Checkpoint to correctly rendering Donkey Kong background

+++ TileEngine/Color.cs
@@ -8,11 +8,11 @@ namespace TileEngine
 {
     class Color
     {
-        public int R { get; set; }
-        public int G { get; set; }
-        public int B { get; set; }
+        public byte R { get; set; }
+        public byte G { get; set; }
+        public byte B { get; set; }
 
-        public Color(int r, int g, int b)
+        public Color(byte r, byte g, byte b)
         {
             R = r;
             G = g;
